rag-project/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-pinecone==0.0.1
langchain-community==0.0.13
pinecone-client==3.0.1
python-dotenv==1.0.1
requests==2.31.0
schedule==1.2.1 